refactor: extract headline to banner component and adjust styles
