+++ src/planets/About.jsx
@@ -2,44 +2,14 @@ import React, { useEffect, useRef, useState } from "react";
 import styles from "./about.module.css";
 import Pillars from "../moons/Pillars.jsx";
 import PillarsSmallScreen from "../moons/PillarsSmallScreen.jsx";
+import Banner from "../moons/Banner.jsx";
+
 
 function About() {
-  const [isInViewport, setIsInViewport] = useState(false);
-  const aboutHeroDividerRef = useRef(null);
-
-  useEffect(() => {
-    const handleIntersection = (entries) => {
-      const isIntersecting = entries[0].isIntersecting;
-      setIsInViewport(isIntersecting);
-    };
-
-    const observer = new IntersectionObserver(handleIntersection);
-    const currentRef = aboutHeroDividerRef.current;
-
-    if (currentRef) {
-      observer.observe(currentRef);
-    }
-
-    return () => {
-      if (currentRef) {
-        observer.unobserve(currentRef);
-      }
-    };
-  }, []); 
 
   return (
     <div className={styles.about}>
-      <div ref={aboutHeroDividerRef} className={styles.aboutHeroDivider}>
-        <div>
-          <h1
-            className={`${styles.noAboutHeadline} ${
-              isInViewport ? styles.aboutHeadline : styles.noAboutHeadline
-            }`}
-          >
-            About
-          </h1>
-        </div>
-      </div>
+      <Banner name={"About"}/>
       <div className={styles.aboutIntro}>
         <h2> Pillars of Tech Risers Women</h2>
         <p>
